Added and finished Patch and put tests for the cart @see: UpdateCartRequest.java , TestUpdateExistingCart.java -- changed some assertion methods. -- added schema validations for the UpdateCartRequest class using    updateCartSchema.json

+++ src/test/java/utils/validation/Fields/ExtractAllCartItems.java
@@ -13,7 +13,7 @@ import static utils.validation.Fields.StaticFieldValidator.validateInt;
 public class ExtractAllCartItems implements CONSTANT_VALIDATION {
     public Items[] extractAllCarts(Items[] items) {
         for (Items item : items) {
-            validateInt().intGreaterEqualTo(item.getId(), 1);
+            validateInt().intGreaterEqualTo((Integer) item.getId(), 1);
             ids.add(item.getId());
             userIds.add(item.getUserId());
             cartDates.add(item.getDate());
